requirements.txt
pandas
numpy
matplotlib
python-docx
openpyxl